Minor fixes.
Fixed shutdown. Now, when you close an app it tries to finish all the
renders safely, deleteing old placeholders.

+++ src/ru/dolika/fractal/Tiles.java
@@ -7,7 +7,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.attribute.FileAttribute;
 import java.util.concurrent.Semaphore;
 
 import javax.imageio.ImageIO;
@@ -46,20 +45,18 @@ public class Tiles {
 		return img;
 	}
 
-	private BufferedImage getTileImage(final long numX, final long numY,
-			final double zoom) {
-		Runnable generator = new Runnable() {
-			public void run() {
-				generateAndPutTile(numX, numY, zoom);
-			}
+	private BufferedImage getTileImage(final long numX, final long numY, final double zoom) {
+		Runnable generator = () -> {
+			generateAndPutTile(numX, numY, zoom);
 		};
+
 		String outputFileName = tileToFile(numX, numY, zoom);
 		File need = new File(outputFileName);
 		if (!Files.exists(need.toPath())) {
 			Path directory = new File(tileToDirectory(zoom)).toPath();
 			if (!Files.exists(directory)) {
 				try {
-					Files.createDirectories(directory, new FileAttribute[0]);
+					Files.createDirectories(directory);
 				} catch (IOException e) {
 					e.printStackTrace();
 				}
@@ -70,6 +67,7 @@ public class Tiles {
 				// not critical
 			}
 			Thread thread = new Thread(generator);
+			thread.setDaemon(true);
 			thread.setPriority(Thread.MIN_PRIORITY);
 			thread.setName("Generator Z:" + zoom + ";X:" + numX + ";Y:" + numY);
 			thread.start();
@@ -85,46 +83,74 @@ public class Tiles {
 		return img;
 	}
 
-	Semaphore generatorSemaphore = new Semaphore(Runtime.getRuntime()
-			.availableProcessors() * 2, false);
+	Semaphore generatorSemaphore = new Semaphore(Runtime.getRuntime().availableProcessors() * 2, false);
 
 	private void generateAndPutTile(long numX, long numY, double zoom) {
+
+		final Thread current = Thread.currentThread();
+
+		Thread shutdownHook = new Thread(() -> {
+			long time = System.currentTimeMillis();
+			System.out.println(current.getName() + " interrupting generation loop");
+			current.interrupt();
+			System.out.println(current.getName() + " waiting to finish");
+			try {
+				current.join();
+			} catch (InterruptedException e) {
+				e.printStackTrace();
+			}
+			System.out.println(
+					current.getName() + " finished in " + (System.currentTimeMillis() - time) + "ms after interrupt");
+
+		});
+
+		Runtime.getRuntime().addShutdownHook(shutdownHook);
+
+		final String outputFileName = tileToFile(numX, numY, zoom);
+
 		try {
 			generatorSemaphore.acquire();
+		} catch (InterruptedException e) {
+			// if we are interrupted, delete placeholder
 			try {
-				BufferedImage img = generateTile(numX, numY, zoom);
-
-				String outputFileName = tileToFile(numX, numY, zoom);
-
-				try {
-
-					ImageIO.write(img, "PNG", new File(outputFileName));
-				} catch (IOException e) {
-					try {
-						Files.delete(new File(outputFileName).toPath());
-					} catch (IOException e1) {
-						e1.printStackTrace();
-					}
-				}
-			} finally {
-				generatorSemaphore.release();
+				Files.delete(new File(outputFileName).toPath());
+			} catch (IOException e1) {
 			}
-		} catch (InterruptedException e1) {
-			e1.printStackTrace();
+			System.err.println(
+					Thread.currentThread().getName() + " interrupted before generation, so placeholder is deleted");
+
+			try {
+				Runtime.getRuntime().removeShutdownHook(shutdownHook);
+			} catch (IllegalStateException eeill) {
+			}
+			return;
+		}
+		try {
+			BufferedImage img = generateTile(numX, numY, zoom);
+			if (img == null) {
+				System.err.println(current.getName() + " deleting placeholder");
+				Files.delete(new File(outputFileName).toPath());
+				return;
+			}
+			// overwrite placeholder image with newly generated one
+			ImageIO.write(img, "PNG", new File(outputFileName));
+		} catch (IOException e) {
+			// if failed - try to delete placeholder
+			try {
+				Files.delete(new File(outputFileName).toPath());
+			} catch (IOException e1) {
+			}
+		} finally {
+			generatorSemaphore.release();
+		}
+		try {
+			Runtime.getRuntime().removeShutdownHook(shutdownHook);
+		} catch (IllegalStateException eeill) {
 		}
 	}
 
-	private String tileToFile(long numX, long numY, double zoom) {
-		return tileToDirectory(zoom) + "tile-" + numX + "-" + numY + ".png";
-	}
-
-	private String tileToDirectory(double zoom) {
-		return ".fractal/cache/" + zoom + "/";
-	}
-
 	private BufferedImage generateTile(long numX, long numY, double zoom) {
-		BufferedImage img = new BufferedImage(tileDim, tileDim,
-				BufferedImage.TYPE_INT_RGB);
+		BufferedImage img = new BufferedImage(tileDim, tileDim, BufferedImage.TYPE_INT_RGB);
 		for (int x = 0; x < tileDim; x++) {
 			for (int y = 0; y < tileDim; y++) {
 
@@ -143,6 +169,11 @@ public class Tiles {
 					Zr = Zrnew + cr;
 					Zi = Zinew + ci;
 					i++;
+
+					if (Thread.interrupted()) {
+						System.err.println(Thread.currentThread().getName() + " interrupted during generation process");
+						return null;
+					}
 				}
 
 				if (((Zr * Zr) + (Zi * Zi)) >= 4) {
@@ -155,22 +186,27 @@ public class Tiles {
 		return img;
 	}
 
+	private String tileToFile(long numX, long numY, double zoom) {
+		return tileToDirectory(zoom) + "tile-" + numX + "-" + numY + ".png";
+	}
+
+	private String tileToDirectory(double zoom) {
+		return ".fractal/cache/" + zoom + "/";
+	}
+
 	private BufferedImage _emptyTile = null;
 
 	private BufferedImage getEmptyTile() {
 		if (_emptyTile == null) {
-			_emptyTile = new BufferedImage(tileDim, tileDim,
-					BufferedImage.TYPE_INT_RGB);
+			_emptyTile = new BufferedImage(tileDim, tileDim, BufferedImage.TYPE_INT_RGB);
 			Graphics g = _emptyTile.getGraphics();
 			g.setColor(Color.white);
 			g.fillRect(0, 0, tileDim, tileDim);
 			g.setColor(Color.BLACK);
-			g.drawLine(_emptyTile.getWidth() / 4, _emptyTile.getWidth() / 4,
-					_emptyTile.getWidth() * 3 / 4,
+			g.drawLine(_emptyTile.getWidth() / 4, _emptyTile.getWidth() / 4, _emptyTile.getWidth() * 3 / 4,
 					_emptyTile.getWidth() * 3 / 4);
-			g.drawLine(_emptyTile.getWidth() / 4,
-					_emptyTile.getWidth() * 3 / 4,
-					_emptyTile.getWidth() * 3 / 4, _emptyTile.getWidth() / 4);
+			g.drawLine(_emptyTile.getWidth() / 4, _emptyTile.getWidth() * 3 / 4, _emptyTile.getWidth() * 3 / 4,
+					_emptyTile.getWidth() / 4);
 		}
 		return _emptyTile;
 	}
